# apps/server/docs/openapi.yaml
components:
  schemas:
    AuthenticatedUserData:
      description: The signed-in user, carrying only what a client may see.
      properties:
        email:
          title: Email
          type: string
        id:
          format: uuid4
          title: Id
          type: string
      required:
      - id
      - email
      title: AuthenticatedUserData
      type: object
    AuthenticatedUserResponse:
      description: The `{ data }` envelope every successful auth route answers with.
      properties:
        data:
          $ref: '#/components/schemas/AuthenticatedUserData'
      required:
      - data
      title: AuthenticatedUserResponse
      type: object
    ChangePasswordRequest:
      additionalProperties: false
      description: The body of a password change, which proves the caller knows the current password.
      properties:
        current_password:
          maxLength: 72
          title: Current Password
          type: string
        new_password:
          maxLength: 72
          minLength: 8
          title: New Password
          type: string
      required:
      - current_password
      - new_password
      title: ChangePasswordRequest
      type: object
    ErrorCode:
      description: Every error code this application answers with, namespaced DOMAIN_REASON.
      enum:
      - AUTH_ADMIN_REQUIRED
      - AUTH_INVALID_CREDENTIALS
      - AUTH_REQUIRED
      - AUTH_SESSION_EXPIRED
      - AUTH_EMAIL_ALREADY_REGISTERED
      - AUTH_RESET_TOKEN_INVALID
      - CSRF_HEADER_MISSING
      - IDEMPOTENCY_KEY_REUSED
      - INPUT_PAYLOAD_TOO_LARGE
      - INPUT_VALIDATION_ERROR
      - RATE_LIMIT_EXCEEDED
      - ROUTING_METHOD_NOT_ALLOWED
      - ROUTING_NOT_FOUND
      - SERVER_DATABASE_UNAVAILABLE
      - SERVER_INTERNAL_ERROR
      - SERVER_RATE_LIMIT_UNAVAILABLE
      - SERVER_REQUEST_TIMEOUT
      title: ErrorCode
      type: string
    ErrorResponse:
      description: 'One failed request: a registry code the client switches on and a human-readable message.


        `field_errors` is populated only for `INPUT_VALIDATION_ERROR` and omitted from the body

        entirely otherwise, so every other failure keeps the two-key shape clients already handle. It

        exists because a form has to show a message beside the input it belongs to, and the prose

        `error` string is documented as never parsed; reading the field out of it would make a

        human-readable sentence into an interface.'
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
          description: Machine-readable code from the error registry.
        error:
          description: Human-readable message; never parsed by clients.
          title: Error
          type: string
        field_errors:
          anyOf:
          - items:
              $ref: '#/components/schemas/FieldError'
            type: array
          - type: 'null'
          description: Per-field validation failures, present only for a 400.
          title: Field Errors
      required:
      - code
      - error
      title: ErrorResponse
      type: object
    FieldError:
      description: One field of a request body that failed validation, and why.
      properties:
        field:
          description: The offending field, as a dotted path without the body prefix.
          title: Field
          type: string
        message:
          description: Why this field was rejected.
          title: Message
          type: string
      required:
      - field
      - message
      title: FieldError
      type: object
    ForgotPasswordData:
      description: The confirmation a reset request answers with, identical for every address.
      properties:
        message:
          title: Message
          type: string
      required:
      - message
      title: ForgotPasswordData
      type: object
    ForgotPasswordRequest:
      additionalProperties: false
      description: 'The body of a reset request.


        The email carries no pattern, for the reason `LoginRequest` gives: the answer is the same 200

        for every address, and a 400 for one that is not even registrable would be the one response

        that differed.'
      properties:
        email:
          maxLength: 320
          minLength: 1
          title: Email
          type: string
      required:
      - email
      title: ForgotPasswordRequest
      type: object
    ForgotPasswordResponse:
      description: The `{ data }` envelope of a reset request.
      properties:
        data:
          $ref: '#/components/schemas/ForgotPasswordData'
      required:
      - data
      title: ForgotPasswordResponse
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    HealthLiveness:
      description: 'Body of `GET /health`: the process is up.'
      properties:
        status:
          const: ok
          title: Status
          type: string
      required:
      - status
      title: HealthLiveness
      type: object
    HealthReadiness:
      description: 'Body of `GET /health/ready`: whether Postgres answered within the deadline.'
      properties:
        db:
          enum:
          - connected
          - disconnected
          title: Db
          type: string
        status:
          enum:
          - ok
          - degraded
          title: Status
          type: string
      required:
      - status
      - db
      title: HealthReadiness
      type: object
    LoginRequest:
      additionalProperties: false
      description: 'The body of a sign-in.


        The email carries no pattern. A malformed address is a failed sign-in rather than a malformed

        request, and answering 400 here would tell a caller that an address is not even registrable,

        which is one more bit than a sign-in should reveal.'
      properties:
        email:
          maxLength: 320
          title: Email
          type: string
        password:
          maxLength: 72
          title: Password
          type: string
      required:
      - email
      - password
      title: LoginRequest
      type: object
    RegisterRequest:
      additionalProperties: false
      description: The body of a registration.
      properties:
        email:
          maxLength: 320
          pattern: ^[^@\s]+@[^@\s.]+(\.[^@\s.]+)+$
          title: Email
          type: string
        password:
          maxLength: 72
          minLength: 8
          title: Password
          type: string
      required:
      - email
      - password
      title: RegisterRequest
      type: object
    ResetPasswordRequest:
      additionalProperties: false
      description: 'The body of a reset: the token from the email''s link and the new password.


        The password takes registration''s constraints, since it becomes the account''s password on

        exactly the same terms. The token''s ceiling is generous against the 43 characters a real one

        has, and exists so an oversized value is refused before it is hashed.'
      properties:
        password:
          maxLength: 72
          minLength: 8
          title: Password
          type: string
        token:
          maxLength: 256
          minLength: 1
          title: Token
          type: string
      required:
      - token
      - password
      title: ResetPasswordRequest
      type: object
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  title: template-fastapi-nuxt
  version: 0.1.0
openapi: 3.1.0
paths:
  /health:
    get:
      description: Answer 200 without touching any dependency.
      operationId: read_liveness_health_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthLiveness'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The request failed validation
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: An unexpected error occurred
      summary: Read Liveness
      tags:
      - health
  /health/ready:
    get:
      description: Answer 200 when Postgres answers within the deadline, and 503 otherwise.
      operationId: read_readiness_health_ready_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthReadiness'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The request failed validation
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: An unexpected error occurred
        '503':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthReadiness'
          description: Postgres did not answer within the deadline
      summary: Read Readiness
      tags:
      - health
  /v1/auth/forgot-password:
    post:
      description: 'Enqueue the reset email and answer 200, whether or not an account uses the address.


        The route never looks the address up. The job does, in the worker, so the response and the

        work this request causes are identical for a known and an unknown address (B-14).'
      operationId: request_password_reset_v1_auth_forgot_password_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ForgotPasswordResponse'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The request failed validation
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: An unexpected error occurred
      summary: Request Password Reset
      tags:
      - auth
  /v1/auth/login:
    post:
      description: Open a session for an existing account and set the session cookie.
      operationId: login_v1_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticatedUserResponse'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The request failed validation
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: An unexpected error occurred
      summary: Login
      tags:
      - auth
  /v1/auth/logout:
    post:
      description: 'End the session if there is one, and clear the cookie either way.


        No injected `Response` is declared. FastAPI assigns a returned `Response` instance directly

        and never merges the injected one''s headers, so anything written to it here would be dropped

        without a test noticing, on the one route whose whole job is to get a cookie header right.'
      operationId: logout_v1_auth_logout_post
      responses:
        '204':
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The request failed validation
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: An unexpected error occurred
      summary: Logout
      tags:
      - auth
  /v1/auth/me:
    get:
      description: Answer the signed-in user's identity, and 401 without a usable session.
      operationId: read_me_v1_auth_me_get
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticatedUserResponse'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The request failed validation
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: An unexpected error occurred
      summary: Read Me
      tags:
      - auth
    patch:
      description: Replace the password after proving the current one, signing out every other session.
      operationId: change_my_password_v1_auth_me_patch
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticatedUserResponse'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The request failed validation
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: An unexpected error occurred
      summary: Change My Password
      tags:
      - auth
  /v1/auth/register:
    post:
      description: Create an account, sign it in, and set the session cookie.
      operationId: register_v1_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticatedUserResponse'
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The request failed validation
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: An unexpected error occurred
      summary: Register
      tags:
      - auth
  /v1/auth/reset-password:
    post:
      description: Set a new password from an emailed token and sign the account out everywhere.
      operationId: reset_password_with_token_v1_auth_reset_password_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
        required: true
      responses:
        '204':
          description: Successful Response
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: The request failed validation
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: An unexpected error occurred
      summary: Reset Password With Token
      tags:
      - auth